Properly map user/group if entry with same name exists in container
With `--map-user` (the default), the a user and group of the same name
as the host was unconditionally created inside the container. This
created problems if the calling user had the same user name or group
name as a group which already existed inside the container. This change
checks for a same-named user or group before attempting to create it
inside the container.